island: move the glasses device-state store into island (+ wifi status facade)
The glasses store is the source of device state (connection, wifi, battery, OTA,
hotspot) that every device facade needs, so it moves into island — and the wifi
status read-model that was blocked on it now lands too.

- stores/glasses.ts -> modules/island/src/stores/glasses.ts. btsdk type imports
  switch to the relative _internal path (standalone-build-safe, like the
  passthrough); connection predicates come from island's GlassesReadiness.
- @/stores/glasses becomes a thin re-export shim — the ~46 existing importers are
  unchanged (the escape hatch from the Phase-1 plan). Also surfaced as
  island.glassesStore (the raw store) for the app to keep using directly.
- jest: the @mentra/island mock provides the REAL store via jest.requireActual so
  every consumer's tests keep their real setState/getState/subscribe behavior.
- island.glasses.wifi gains status() / onStatus() reading the island-owned store
  — the wifi facade is now complete.

Verified: typecheck 0 new errors across all 46 consumers; jest 265 pass (the real
store runs in-test via requireActual). The island standalone build uses the proven
relative-_internal pattern; CI confirms it (local build is blocked only by the
unrelated cloud-v2 zod dep not being installed in this checkout).
